requirements.txt

flask
numpy
scipy
matplotlib
pycryptodome
pandas
sqlalchemy
